Include conditional for item_orders

+++ app/controllers/merchant/items_controller.rb
@@ -48,8 +48,12 @@ class Merchant::ItemsController < Merchant::BaseController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    flash[:success] = "Yer item walked the plank fer good!"
+    if item.orders.empty?
+      item.destroy
+      flash[:success] = "Yer item walked the plank fer good!"
+    else
+      flash[:failure] = "ARGH! Ye shan't be deletin booty that's already been plundered!"
+    end
     redirect_to '/merchant/items'
   end
